Don't crash if a local stack branch already exists.
Sometimes when a script fails we end up with local branches without the
corresponding remote branch. That causes the script to fail next time
it's invoked when it tries to create a branch with the same name. This
PR fixes it.
Also, this PR makes the name picking less conservative so that branch
names like `username/issue/765` are not considered related to any stack.