Allow the default reviewers to be set via an environment variable.
Users can opt to specify the default reviewer across PRs by setting the
STACK_PR_DEFAULT_REVIEWER variable. The use case is that people
within a group want to assign the group as a reviewer more often than
not. This is a purely opt in if you set the
STACK_PR_DEFAULT_REVIEWER env variable.