Be more conservative when we advance 'main' to 'origin/main'.
We should be doing that only if 'head' contains all commits from 'main'
to 'origin/main' (which will be the case when, e.g., we land a part of
the stack). But in other cases we should keep 'main' intact - otherwise
it will no longer be an ancestor of 'head'.