Delete local stack branches.
The branches `stack-pr` creates are implementation details and should
not be used by the users. However, we kept them locally and it often
created confusion.

With this PR we now delete the branches after the stack is exported.
When we land a stack we re-create the local branches as needed and clean
them up in the end.