Pick a better default for 'base' when it's not specified explicitly.
Previously we were using 'main' as the default, but it's better to use
merge-base between `origin/main` and `head`.

E.g. in the following scenario using 'main' as the base would result in
including commit B into the stack. This is unlikely what the user wants
because the commit B is already landed in remote 'main':

```
 --> a ----> b  ----> c ----> d
   (main)     \          (origin/main)
               \
                ---> e ----> f
                           (head)
```