Remove shell parameter from subprocess utilities
The interface as specified encouraged unsafe uses. Putting two separate
list items as arguments to a run_shell_command with shell=True did not
necessarily guarantee the command receives two arguments, for example,
and command injection was possible. These are not likely to be big
problems for us, but it seems like we should default to being safe.

The removal of shell=True also led to run_chained_commands not working
any more, as that relied on a "&&" argument chaining commands. It would
be possible to implement run_chained_commands in terms of multiple
run_shell_command calls, but combining the return results would have
been tedious. It is easier to change the call sites to use multiple
calls to run_shell_command.

I have been running with these changes for a few days and have not
noticed any failures, and I searched the code base and didn't find any
obvious other places where we might encounter failures. If there are
any, I expect they should be limited in scope and easy to fix.