Automatically rebase the rest of the stack after landing the bottom PR.
When we added required checks to our CI, we lost the ability to land
whole PRs stack at once: we could land the bottom PR just fine, but
before we could land the rest we needed to rebase them - that triggered
CI runs and PRs got blocked by that.

Because of that, we disabled the `land` command on stacks with more
than 1 PR (since we couldn't land them anyway and trying to land more
than one resulted in an error).

However, while we can't land the whole stack, there is no reason for us
to not be able to handle stacks with more than one PR (except that I
couldn't find time to actually do that).

This PR allows us to use `stack-pr land` on stacks with more than one PR
again: it lands the bottommost PR and rebases the rest, so they are
ready to be landed as soon as CI finishes.