Emit terminal hyperlinks
Implements enough of the [terminal hyperlink specification][1] to allow
us to include links to PRs, where they exist, in the terminal output.

This specification is implemented by several terminals, including:
- kitty
- wezterm
- iTerm2
- Alacritty
- Windows Terminal
- Konsole
- xterm.js (VS Code's backing terminal implementation)

See [this tracking repository][2] for a full listing.

The ECMA-48 specification requires that terminals that do not understand
the OSC-8 command code silently ignore it; in compliant terminals that
do not implement the hyperlink specification, users will simply observe
no changes to the output of `stack-pr`. For cases where outputting these
escape codes is still undesirable I have added a `--no-hyperlinks` flag.

[1]: https://gist.github.com/egmontkob/eb114294efbcd5adb1944c9f3cb5feda
[2]: https://github.com/Alhadis/OSC8-Adoption/