Don't attempt to delete remote branch after a PR has landed. (#42)
GitHub deletes the remote branch automatically when the PR is merged, so
we don't to explicitly do that. Trying to delete it currently results in
an error like this:
```
error: unable to delete 'ZolotukhinM/stack/1': remote ref does not exist
error: failed to push some refs to '...'
```
The code for deleting the branch has been sitting there for a long time,
but until a fix in #27 it used incorrect branch names.